+++ EstudosJava/src/estudo_heranca/Motocicleta.java
@@ -3,9 +3,9 @@ package estudo_heranca;
 
 public class Motocicleta extends Automovel{
     @Override
-    public void acelerar(int giradaAcelerador){
-        double porcentagemGirada = giradaAcelerador/100;
-        double velocidade = super.getVelocidade();
+    public void acelerar(float giradaAcelerador){
+        float porcentagemGirada = giradaAcelerador/100;
+        float velocidade = super.getVelocidade();
         velocidade += velocidade * porcentagemGirada; 
         
         super.setVelocidade(velocidade);
